refactor(issue): move new issue to separate file
